pyproject.toml
[project]
name = "data-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastfeedparser>=0.4.4",
    "google-cloud-bigquery>=3.38.0",
    "pandas>=2.3.3",
    "pandas-gbq>=0.31.0",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "yfinance>=0.2.66",
]
